+++ CMakeLists.txt
@@ -1,11 +1,13 @@
-cmake_minimum_required(VERSION 3.10)
+cmake_minimum_required(VERSION 3.20)
 
-project(homeworks)
+project(homeworks LANGUAGES CXX)
+
+set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_FLAGS "-Wall -Wextra -pedantic -std=c++20 -O2")
 
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/lib)
-
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/sandbox)
-
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/additional_tasks)
 
 file(GLOB_RECURSE tasks_dirs LIST_DIRECTORIES true ".")
