+++ additional_tasks/graph/CMakeLists.txt
@@ -23,7 +23,6 @@ add_executable(${PROJECT_NAME} ${source_list} ${lib_source_list})
 # Locate GTest
 enable_testing()
 find_package(GTest REQUIRED)
-
 include_directories(${GTEST_INCLUDE_DIRS})
 
 find_library(Utils ../)
@@ -34,6 +33,7 @@ add_executable(${PROJECT_NAME}_tests ${test_source_list})
 target_link_libraries(
   ${PROJECT_NAME}_tests
   GTest::gtest_main
+  Utils
 )
 
 include(GoogleTest)
