+++ additional_tasks/template_task/CMakeLists.txt
@@ -23,7 +23,6 @@ add_executable(${PROJECT_NAME} ${source_list} ${lib_source_list})
 # Locate GTest
 enable_testing()
 find_package(GTest REQUIRED)
-
 include_directories(${GTEST_INCLUDE_DIRS})
 
 find_library(Utils ../)
